+++ ext/stub/19__closure.zep.c
@@ -12,9 +12,9 @@
 #include <Zend/zend_interfaces.h>
 
 #include "kernel/main.h"
-#include "kernel/fcall.h"
-#include "kernel/operators.h"
+#include "kernel/concat.h"
 #include "kernel/memory.h"
+#include "kernel/fcall.h"
 #include "kernel/object.h"
 
 
@@ -22,6 +22,7 @@ ZEPHIR_INIT_CLASS(stub_19__closure)
 {
 	ZEPHIR_REGISTER_CLASS(stub, 19__closure, stub, 19__closure, stub_19__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
+	zend_declare_property_null(stub_19__closure_ce, SL("name"), ZEND_ACC_PUBLIC|ZEND_ACC_STATIC);
 	return SUCCESS;
 }
 
@@ -29,20 +30,18 @@ PHP_METHOD(stub_19__closure, __invoke)
 {
 	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
 	zend_long ZEPHIR_LAST_CALL_STATUS;
-	zval *matches_param = NULL;
-	zval matches;
+	zval name, _0;
 	zval *this_ptr = getThis();
 
-	ZVAL_UNDEF(&matches);
-	ZEND_PARSE_PARAMETERS_START(1, 1)
-		ZEPHIR_Z_PARAM_ARRAY(matches, matches_param)
-	ZEND_PARSE_PARAMETERS_END();
+	ZVAL_UNDEF(&name);
+	ZVAL_UNDEF(&_0);
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
-	zephir_fetch_params(1, 1, 0, &matches_param);
-	zephir_get_arrval(&matches, matches_param);
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "issue2321dourlencode", NULL, 0, &matches);
+	zephir_read_static_property_ce(&name, stub_19__closure_ce, SL("name"), PH_NOISY_CC);
+
+	ZEPHIR_CALL_METHOD(&_0, this_ptr, "issue2497helper", NULL, 0);
 	zephir_check_call_status();
+	ZEPHIR_CONCAT_VSV(return_value, &_0, ":", &name);
 	RETURN_MM();
 }