+++ ext/stub/20__closure.zep.c
@@ -41,7 +41,7 @@ PHP_METHOD(stub_20__closure, __invoke)
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
 	zephir_fetch_params(1, 1, 0, &matches_param);
 	zephir_get_arrval(&matches, matches_param);
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "issue2321protectedupper", NULL, 0, &matches);
+	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "issue2321dourlencode", NULL, 0, &matches);
 	zephir_check_call_status();
 	RETURN_MM();
 }