+++ ext/stub/23__closure.zep.h
@@ -5,9 +5,8 @@ ZEPHIR_INIT_CLASS(stub_23__closure);
 
 PHP_METHOD(stub_23__closure, __invoke);
 
-ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_23__closure___invoke, 0, 0, 2)
-	ZEND_ARG_INFO(0, carry)
-	ZEND_ARG_INFO(0, row)
+ZEND_BEGIN_ARG_INFO_EX(arginfo_stub_23__closure___invoke, 0, 0, 1)
+	ZEND_ARG_INFO(0, x)
 ZEND_END_ARG_INFO()
 
 ZEPHIR_INIT_FUNCS(stub_23__closure_method_entry) {