+++ ext/stub/issue2167.zep.c
@@ -128,7 +128,7 @@ PHP_METHOD(Stub_Issue2167, reduceWithClosureWrapper)
 	zephir_read_property(&_0, this_ptr, ZEND_STRL("a"), PH_NOISY_CC | PH_READONLY);
 	ZEPHIR_INIT_VAR(&_1);
 	ZEPHIR_INIT_NVAR(&_1);
-	zephir_create_closure_ex(&_1, NULL, stub_23__closure_ce, SL("__invoke"));
+	zephir_create_closure_ex(&_1, NULL, stub_24__closure_ce, SL("__invoke"));
 	ZEPHIR_INIT_VAR(&_2);
 	ZVAL_STRING(&_2, "");
 	ZEPHIR_CALL_FUNCTION(&_3, "array_reduce", NULL, 63, &_0, &_1, &_2);