+++ ext/stub/mcall.zep.c
@@ -847,7 +847,7 @@ PHP_METHOD(Stub_Mcall, testCallablePass)
 
 	ZEPHIR_INIT_VAR(&_0);
 	ZEPHIR_INIT_NVAR(&_0);
-	zephir_create_closure_ex(&_0, NULL, stub_28__closure_ce, SL("__invoke"));
+	zephir_create_closure_ex(&_0, NULL, stub_29__closure_ce, SL("__invoke"));
 	ZEPHIR_RETURN_CALL_METHOD(&a, "setcallable", NULL, 70, &_0);
 	zephir_check_call_status();
 	RETURN_MM();